+++ src/components/index.js
@@ -1,3 +1,4 @@
 export { default as ClickOutside } from './ClickOutside'
+export { default as FocusTrap } from './FocusTrap'
 export { default as MergeRefs } from './MergeRefs'
 export { default as Responsive } from './Responsive'