+++ src/core/index.js
@@ -1,4 +1,5 @@
 export { default as getDisplayName } from './getDisplayName'
 export { default as getScreenSize } from './getScreenSize'
 export { default as minWidthModifier } from './minWidthModifier'
+export { default as setRefs } from './setRefs'
 export { default as watchResize } from './watchResize'